+++ src/@sdk/mutations/types/CreateUserAddress.ts
@@ -2,7 +2,7 @@
 /* eslint-disable */
 // This file was automatically generated and should not be edited.
 
-import { AddressInput } from "./../../../../types/globalTypes";
+import { AddressInput } from "./../../types/globalTypes";
 
 // ====================================================
 // GraphQL mutation operation: CreateUserAddress
@@ -52,6 +52,14 @@ export interface CreateUserAddress_accountAddressCreate_user_defaultShippingAddr
   country: CreateUserAddress_accountAddressCreate_user_defaultShippingAddress_country;
   countryArea: string;
   phone: string | null;
+  /**
+   * Address is user's default billing address.
+   */
+  isDefaultBillingAddress: boolean | null;
+  /**
+   * Address is user's default shipping address.
+   */
+  isDefaultShippingAddress: boolean | null;
 }
 
 export interface CreateUserAddress_accountAddressCreate_user_defaultBillingAddress_country {
@@ -85,6 +93,14 @@ export interface CreateUserAddress_accountAddressCreate_user_defaultBillingAddre
   country: CreateUserAddress_accountAddressCreate_user_defaultBillingAddress_country;
   countryArea: string;
   phone: string | null;
+  /**
+   * Address is user's default billing address.
+   */
+  isDefaultBillingAddress: boolean | null;
+  /**
+   * Address is user's default shipping address.
+   */
+  isDefaultShippingAddress: boolean | null;
 }
 
 export interface CreateUserAddress_accountAddressCreate_user_addresses_country {
@@ -118,6 +134,14 @@ export interface CreateUserAddress_accountAddressCreate_user_addresses {
   country: CreateUserAddress_accountAddressCreate_user_addresses_country;
   countryArea: string;
   phone: string | null;
+  /**
+   * Address is user's default billing address.
+   */
+  isDefaultBillingAddress: boolean | null;
+  /**
+   * Address is user's default shipping address.
+   */
+  isDefaultShippingAddress: boolean | null;
 }
 
 export interface CreateUserAddress_accountAddressCreate_user {