+++ src/@sdk/mutations/types/DeleteUserAddress.ts
@@ -50,6 +50,14 @@ export interface DeleteUserAddress_accountAddressDelete_user_defaultShippingAddr
   country: DeleteUserAddress_accountAddressDelete_user_defaultShippingAddress_country;
   countryArea: string;
   phone: string | null;
+  /**
+   * Address is user's default billing address.
+   */
+  isDefaultBillingAddress: boolean | null;
+  /**
+   * Address is user's default shipping address.
+   */
+  isDefaultShippingAddress: boolean | null;
 }
 
 export interface DeleteUserAddress_accountAddressDelete_user_defaultBillingAddress_country {
@@ -83,6 +91,14 @@ export interface DeleteUserAddress_accountAddressDelete_user_defaultBillingAddre
   country: DeleteUserAddress_accountAddressDelete_user_defaultBillingAddress_country;
   countryArea: string;
   phone: string | null;
+  /**
+   * Address is user's default billing address.
+   */
+  isDefaultBillingAddress: boolean | null;
+  /**
+   * Address is user's default shipping address.
+   */
+  isDefaultShippingAddress: boolean | null;
 }
 
 export interface DeleteUserAddress_accountAddressDelete_user_addresses_country {
@@ -116,6 +132,14 @@ export interface DeleteUserAddress_accountAddressDelete_user_addresses {
   country: DeleteUserAddress_accountAddressDelete_user_addresses_country;
   countryArea: string;
   phone: string | null;
+  /**
+   * Address is user's default billing address.
+   */
+  isDefaultBillingAddress: boolean | null;
+  /**
+   * Address is user's default shipping address.
+   */
+  isDefaultShippingAddress: boolean | null;
 }
 
 export interface DeleteUserAddress_accountAddressDelete_user {