+++ src/@sdk/queries/types/ProductDetails.ts
@@ -506,7 +506,7 @@ export interface ProductDetails_product_category_products_edges_node {
    * Lists the storefront product's pricing, the current price and discounts, only meant for displaying.
    */
   pricing: ProductDetails_product_category_products_edges_node_pricing | null;
-  category: ProductDetails_product_category_products_edges_node_category;
+  category: ProductDetails_product_category_products_edges_node_category | null;
 }
 
 export interface ProductDetails_product_category_products_edges {
@@ -545,6 +545,42 @@ export interface ProductDetails_product_images {
    * The URL of the image.
    */
   url: string;
+}
+
+export interface ProductDetails_product_attributes_attribute {
+  __typename: "Attribute";
+  /**
+   * The ID of the object.
+   */
+  id: string;
+  /**
+   * Name of an attribute displayed in the interface.
+   */
+  name: string | null;
+}
+
+export interface ProductDetails_product_attributes_values {
+  __typename: "AttributeValue";
+  /**
+   * The ID of the object.
+   */
+  id: string;
+  /**
+   * Name of a value displayed in the interface.
+   */
+  name: string | null;
+}
+
+export interface ProductDetails_product_attributes {
+  __typename: "SelectedAttribute";
+  /**
+   * Name of an attribute displayed in the interface.
+   */
+  attribute: ProductDetails_product_attributes_attribute;
+  /**
+   * Values of an attribute.
+   */
+  values: (ProductDetails_product_attributes_values | null)[];
 }
 
 export interface ProductDetails_product_variants_pricing_priceUndiscounted_gross {
@@ -663,7 +699,7 @@ export interface ProductDetails_product_variants_attributes_attribute {
   name: string | null;
 }
 
-export interface ProductDetails_product_variants_attributes_value {
+export interface ProductDetails_product_variants_attributes_values {
   __typename: "AttributeValue";
   /**
    * The ID of the object.
@@ -686,9 +722,9 @@ export interface ProductDetails_product_variants_attributes {
    */
   attribute: ProductDetails_product_variants_attributes_attribute;
   /**
-   * The value or the first value of an attribute.
+   * Values of an attribute.
    */
-  value: ProductDetails_product_variants_attributes_value | null;
+  values: (ProductDetails_product_variants_attributes_values | null)[];
 }
 
 export interface ProductDetails_product_variants {
@@ -717,14 +753,6 @@ export interface ProductDetails_product_variants {
   attributes: ProductDetails_product_variants_attributes[];
 }
 
-export interface ProductDetails_product_availability {
-  __typename: "ProductPricingInfo";
-  /**
-   * Whether it is in stock and visible or not.
-   */
-  available: boolean | null;
-}
-
 export interface ProductDetails_product {
   __typename: "Product";
   /**
@@ -745,11 +773,15 @@ export interface ProductDetails_product {
    */
   pricing: ProductDetails_product_pricing | null;
   descriptionJson: any;
-  category: ProductDetails_product_category;
+  category: ProductDetails_product_category | null;
   /**
    * List of images for the product.
    */
   images: (ProductDetails_product_images | null)[] | null;
+  /**
+   * List of attributes assigned to this product.
+   */
+  attributes: ProductDetails_product_attributes[];
   /**
    * List of variants for the product.
    */
@@ -757,9 +789,9 @@ export interface ProductDetails_product {
   seoDescription: string | null;
   seoTitle: string | null;
   /**
-   * Informs about product's availability in the storefront, current price and discounts.
+   * Whether the product is in stock and visible or not.
    */
-  availability: ProductDetails_product_availability | null;
+  isAvailable: boolean | null;
 }
 
 export interface ProductDetails {