+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable, unreleased changes to this project will be documented in this file.
 - Account confirmation mechanism - #565 by @tomaszszymanski129
 - Add missing product attributes on product page - #536 by @orzechdev
 - Change register mutation to accountRegister - #549 by @tomaszszymanski129
+- Add `ProductVariantPicker` component supporting multiple product variant attributes - #550 by @orzechdev
 - Fix not working storefront when no data in saleor database exist - #551 by @orzechdev
 
 ## 0.7.0